Quality of Life Changes
Revamped the model loader to use Ollama api instead of windows shell.
Should be friendly for cross platform.

Made a number of formatting improvements with deepseek-r1 in mind. Some
simple changes to hide "<think></think>" When unused or space it better
when used.

Fixed a bug where clicking in the textbox while the bot is typing caused
the responses to appear at the cursor's index.